CRUD endpoints for User

+++ prmp/Spring Boot Backend/src/main/java/com/prmp/repository/UserRepository.java
@@ -6,8 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.prmp.entity.User;
 
-public interface UserRepository
-        extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     Optional<User> findByEmail(String email);
 }
